Split AppleWin -> Core with the functionality really needed by the emulator.
And a few things that were stricly related (like version, directories).
