+++ source/AY8910.cpp
@@ -28,7 +28,7 @@
 
 #include "AY8910.h"
 
-#include "AppleWin.h"		// For g_fh
+#include "Core.h"		// For g_fh
 #include "YamlHelper.h"
 
 /* The AY white noise RNG algorithm is based on info from MAME's ay8910.c -
