+++ source/CPU.cpp
@@ -87,7 +87,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "StdAfx.h"
 
 #include "CPU.h"
-#include "AppleWin.h"
+#include "Core.h"
 #include "CardManager.h"
 #include "Memory.h"
 #include "Mockingboard.h"
