+++ source/CardManager.cpp
@@ -30,7 +30,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "StdAfx.h"
 
 #include "CardManager.h"
-#include "AppleWin.h"
+#include "Core.h"
 
 #include "Disk.h"
 #include "MouseInterface.h"
