+++ source/Configuration/About.cpp
@@ -24,6 +24,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "StdAfx.h"
 
 #include "About.h"
+#include "../Core.h"
 #include "../AppleWin.h"
 #include "../Windows/WinFrame.h"
 #include "../resource/resource.h"
