+++ source/Debugger/Debug.cpp
@@ -35,6 +35,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "DebugDefs.h"
 
 #include "../AppleWin.h"
+#include "../Core.h"
 #include "../CardManager.h"
 #include "../CPU.h"
 #include "../Disk.h"
