+++ source/Debugger/Debugger_Assembler.cpp
@@ -32,6 +32,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 
 #include "../CPU.h"
 #include "../Memory.h"
+#include "../Windows/WinFrame.h"
 
 #define DEBUG_ASSEMBLER 0
