+++ source/Debugger/Debugger_Display.cpp
@@ -33,6 +33,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "Debugger_Display.h"
 
 #include "../AppleWin.h"
+#include "../Core.h"
 #include "../CPU.h"
 #include "../Frame.h"
 #include "../Windows/WinFrame.h"
