+++ source/Disk.cpp
@@ -35,6 +35,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "SaveState_Structs_v1.h"
 
 #include "AppleWin.h"
+#include "Core.h"
 #include "CPU.h"
 #include "DiskImage.h"
 #include "Log.h"
