+++ source/Harddisk.cpp
@@ -29,6 +29,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "StdAfx.h"
 
 #include "Harddisk.h"
+#include "Core.h"
 #include "AppleWin.h"
 #include "CardManager.h"
 #include "CPU.h"
