+++ source/LanguageCard.cpp
@@ -29,7 +29,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "StdAfx.h"
 
 #include "LanguageCard.h"
-#include "AppleWin.h"
+#include "Core.h"
 #include "CPU.h"		// GH#700
 #include "Log.h"
 #include "Memory.h"
