+++ source/Memory.cpp
@@ -31,7 +31,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "StdAfx.h"
 
 #include "Memory.h"
-#include "AppleWin.h"
+#include "Core.h"
 #include "CardManager.h"
 #include "CPU.h"
 #include "Disk.h"
