+++ source/NTSC.cpp
@@ -22,7 +22,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 // Includes
 	#include "StdAfx.h"
 	#include "NTSC.h"
-	#include "AppleWin.h"
+	#include "Core.h"
 	#include "CPU.h"	// CpuGetCyclesThisVideoFrame()
 	#include "Frame.h"
 	#include "Memory.h" // MemGetMainPtr(), MemGetAuxPtr(), MemGetAnnunciator()
