+++ source/ParallelPrinter.cpp
@@ -29,7 +29,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "StdAfx.h"
 
 #include "ParallelPrinter.h"
-#include "AppleWin.h"
+#include "Core.h"
 #include "Memory.h"
 #include "Registry.h"
 #include "YamlHelper.h"
