+++ source/SerialComms.cpp
@@ -36,7 +36,6 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "StdAfx.h"
 
 #include "SerialComms.h"
-#include "AppleWin.h"
 #include "CPU.h"
 #include "Windows/WinFrame.h"
 #include "Log.h"
