+++ source/SynchronousEventManager.cpp
@@ -42,7 +42,6 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "StdAfx.h"
 
 #include "SynchronousEventManager.h"
-#include "AppleWin.h"
 
 void SynchronousEventManager::Insert(SyncEvent* pNewEvent)
 {
