+++ source/Tape.cpp
@@ -32,7 +32,6 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "StdAfx.h"
 
 #include "Tape.h"
-#include "AppleWin.h"
 #include "Keyboard.h"
 #include "Memory.h"
 #include "Pravets.h"
